Allow MC to post announcements

+++ app/controllers/announcements_controller.rb
@@ -1,28 +1,38 @@
 # frozen_string_literal: true
 
 class AnnouncementsController < ApplicationController
-  load_and_authorize_resource
-
   def index
     @announcements = Announcement.order(created_at: :desc)
   end
 
   def create
-    announcement = Announcement.new announcement_params
+    if current_user.mc
+      announcement = Announcement.new announcement_params
 
-    redirect_to duties_path if announcement.save
+      redirect_to duties_path if announcement.save
+    else
+      redirect_to duties_path
+    end
   end
 
   def update
-    announcement = Announcement.find params[:id]
+    if current_user.mc
+      announcement = Announcement.find params[:id]
 
-    redirect_to duties_path if announcement.update announcement_params
+      redirect_to duties_path if announcement.update announcement_params
+    else
+      redirect_to duties_path
+    end
   end
 
   def destroy
-    announcement = Announcement.find params[:id]
+    if current_user.mc
+      announcement = Announcement.find params[:id]
 
-    redirect_to duties_path if announcement.destroy
+      redirect_to duties_path if announcement.destroy
+    else
+      redirect_to duties_path
+    end
   end
 
   private
